requirements.txt
attrs==20.3.0
Automat==20.2.0
blis==0.7.4
catalogue==1.0.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
constantly==15.1.0
cryptography==3.4.7
cssselect==1.1.0
cycler==0.10.0
cymem==2.0.5
docx2txt==0.8
en-core-web-sm==2.3.1
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
itemadapter==0.2.0
itemloaders==1.0.4
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.0.1
jsonschema==3.2.0
kiwisolver==1.3.1
lxml==4.6.3
MarkupSafe==1.1.1
matplotlib==3.4.1
murmurhash==1.0.5
nltk==3.5
normalise==0.1.8
numpy==1.20.2
packaging==20.9
pandas==1.2.3
parsel==1.6.0
pathy==0.4.0
pdfminer==20191125
pdfminer.six==20200517
pdfplumber==0.5.27
phonenumbers==8.12.20
Pillow==8.1.2
plac==1.1.3
preshed==3.0.5
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
pydantic==1.7.3
PyDispatcher==2.0.5
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyPDF2==1.26.0
pyresparser==1.0.6
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
queuelib==1.5.0
regex==2021.3.17
requests==2.25.1
resume-parser==0.8
roman==3.3
scikit-learn==0.22.2.post1
scipy==1.6.2
Scrapy==2.4.1
service-identity==18.1.0
six==1.15.0
smart-open==3.0.0
sortedcontainers==2.3.0
spacy==2.3.5
spacy-legacy==3.0.2
srsly==1.0.5
stemming==1.0.1
thinc==7.4.5
threadpoolctl==2.1.0
tika==1.24
tqdm==4.59.0
Twisted==21.2.0
typer==0.3.2
urllib3==1.26.4
w3lib==1.22.0
Wand==0.6.6
wasabi==0.8.2
zope.interface==5.3.0

